Rename kwarg/cli arguments and move header logic from Server to Config
As requested in PR #321, I'm renaming the --default-header CLI argument
to --header, and renaming the kwarg/instance variable custom_headers to
headers. Also, I'm moving (most) of the header generation logic from
main.Server to config.Config, so that most of the headers only get
generated once, instead of on every tick. The most computationally
intense generation will still happen ever tick, and that is the
date-stamp generation.