main.py imports code before eventloop starts running
Before this patch there was an inconsistency where if you ran uvicorn
from the CLI versus programmatically the eventloop would either be or
not be running when the code was imported. It is generally good practice
to make sure that the event loop isn't running when code is imported
since async code can only be run at import time using the
loop.run_until_complete(...) command which requires the event loop to
not be running.

This patch makes sure that the code is loaded after the event loop is
configured, but before it starts running. This allows for consistent
behavior across deployment options.