Support disabling default Date header
When an origin server does not have a sufficiently reliable clock,
section 7.1.1.2 of RFC 7231 stipulates that it "MUST NOT" send
a Date header field.

By default, uvicorn adds a Date header to each response. This change
adds the date_header flag to Config, as well as the commandline
option --no-date-header, through which this behaviour can be disabled.