add run_server(server) convenience func
the run() function contains valuable logic for kicking of
the StatReloader and the Multiprocess workers, however it
does not give you access to the server instance before
starting and subsequently blocking. thus splitting out the
server start logic into a function accepting an already
intialized server instance solves this problem.