Add GitHub Actions timeout for tests
With the introduction of platform-specific worker testing, subprocesses,
and operating system signal handling, there is potential for specific
workflow jobs to occasionally hang. In these cases, it can help to set
a timeout on the test step to ensure the workflow jobs don't hang until
the default 360 minute (6 hour) timeout.

https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions